Separated Far's TopologyDescriptor from the RefinerFactory classes:     - moved TopologyRefiner out of the RefinerFactoryBase into Far     - moved implementation of its Factory<MESH> to far/topologyDescriptor.*     - updated examples and tutorials (no more references to FactoryBase)
