+++ opensubdiv/far/topologyRefinerFactory.cpp
@@ -143,7 +143,7 @@ TopologyRefinerFactoryBase::prepareComponentTopologyAssignment(TopologyRefiner& 
     }
 
     //  Now that we have a valid base level, initialize the Refiner's component inventory:
-    initializeBaseInventory(refiner);
+    refiner.initializeInventory();
     return true;
 }
 
@@ -281,143 +281,6 @@ TopologyRefinerFactoryBase::prepareFaceVaryingChannels(TopologyRefiner& refiner)
         baseLevel.completeFVarChannelTopology(channel, regBoundaryValence);
     }
     return true;
-}
-
-
-//
-// Specialization for raw topology data
-//
-template <>
-bool
-TopologyRefinerFactory<TopologyRefinerFactoryBase::TopologyDescriptor>::resizeComponentTopology(
-    TopologyRefiner & refiner, TopologyDescriptor const & desc) {
-
-    setNumBaseVertices(refiner, desc.numVertices);
-    setNumBaseFaces(refiner, desc.numFaces);
-
-    for (int face=0; face<desc.numFaces; ++face) {
-
-        setNumBaseFaceVertices(refiner, face, desc.numVertsPerFace[face]);
-    }
-    return true;
-}
-
-template <>
-bool
-TopologyRefinerFactory<TopologyRefinerFactoryBase::TopologyDescriptor>::assignComponentTopology(
-    TopologyRefiner & refiner, TopologyDescriptor const & desc) {
-
-    for (int face=0, idx=0; face<desc.numFaces; ++face) {
-
-        IndexArray dstFaceVerts = getBaseFaceVertices(refiner, face);
-
-        if (desc.isLeftHanded) {
-            dstFaceVerts[0] = desc.vertIndicesPerFace[idx++];
-            for (int vert=dstFaceVerts.size()-1; vert > 0; --vert) {
-
-                dstFaceVerts[vert] = desc.vertIndicesPerFace[idx++];
-            }
-        } else {
-            for (int vert=0; vert<dstFaceVerts.size(); ++vert) {
-
-                dstFaceVerts[vert] = desc.vertIndicesPerFace[idx++];
-            }
-        }
-    }
-    return true;
-}
-
-template <>
-bool
-TopologyRefinerFactory<TopologyRefinerFactoryBase::TopologyDescriptor>::assignComponentTags(
-    TopologyRefiner & refiner, TopologyDescriptor const & desc) {
-
-    if ((desc.numCreases>0) and desc.creaseVertexIndexPairs and desc.creaseWeights) {
-
-        int const * vertIndexPairs = desc.creaseVertexIndexPairs;
-        for (int edge=0; edge<desc.numCreases; ++edge, vertIndexPairs+=2) {
-
-            Index idx = findBaseEdge(refiner, vertIndexPairs[0], vertIndexPairs[1]);
-
-            if (idx!=Vtr::INDEX_INVALID) {
-                setBaseEdgeSharpness(refiner, idx, desc.creaseWeights[edge]);
-            } else {
-                char msg[1024];
-                snprintf(msg, 1024, "Edge %d specified to be sharp does not exist (%d, %d)",
-                    edge, vertIndexPairs[0], vertIndexPairs[1]);
-                reportInvalidTopology(Vtr::internal::Level::TOPOLOGY_INVALID_CREASE_EDGE, msg, desc);
-            }
-        }
-    }
-
-    if ((desc.numCorners>0) and desc.cornerVertexIndices and desc.cornerWeights) {
-
-        for (int vert=0; vert<desc.numCorners; ++vert) {
-
-            int idx = desc.cornerVertexIndices[vert];
-
-            if (idx >= 0 and idx < getNumBaseVertices(refiner)) {
-                setBaseVertexSharpness(refiner, idx, desc.cornerWeights[vert]);
-            } else {
-                char msg[1024];
-                snprintf(msg, 1024, "Vertex %d specified to be sharp does not exist", idx);
-                reportInvalidTopology(Vtr::internal::Level::TOPOLOGY_INVALID_CREASE_VERT, msg, desc);
-            }
-        }
-    }
-    if (desc.numHoles>0) {
-        for (int i=0; i<desc.numHoles; ++i) {
-            setBaseFaceHole(refiner, desc.holeIndices[i], true);
-        }
-    }
-    return true;
-}
-
-template <>
-bool
-TopologyRefinerFactory<TopologyRefinerFactoryBase::TopologyDescriptor>::assignFaceVaryingTopology(
-    TopologyRefiner & refiner, TopologyDescriptor const & desc) {
-
-    if (desc.numFVarChannels>0) {
-
-        for (int channel=0; channel<desc.numFVarChannels; ++channel) {
-
-            int        numFVarValues = desc.fvarChannels[channel].numValues;
-            int const* srcFVarValues = desc.fvarChannels[channel].valueIndices;
-
-            createBaseFVarChannel(refiner, numFVarValues);
-
-            for (int face = 0, srcNext = 0; face < desc.numFaces; ++face) {
-
-                IndexArray dstFaceFVarValues = getBaseFaceFVarValues(refiner, face, channel);
-
-                if (desc.isLeftHanded) {
-                    dstFaceFVarValues[0] = srcFVarValues[srcNext++];
-                    for (int vert = dstFaceFVarValues.size() - 1; vert > 0; --vert) {
-                        
-                        dstFaceFVarValues[vert] = srcFVarValues[srcNext++];
-                    }
-                } else {
-                    for (int vert = 0; vert < dstFaceFVarValues.size(); ++vert) {
-                        
-                        dstFaceFVarValues[vert] = srcFVarValues[srcNext++];
-                    }
-                }
-            }
-        }
-    }
-    return true;
-}
-
-template <>
-void
-TopologyRefinerFactory<TopologyRefinerFactoryBase::TopologyDescriptor>::reportInvalidTopology(
-    TopologyError /* errCode */, char const * msg, TopologyDescriptor const& /* mesh */) {
-    Warning(msg);
-}
-
-TopologyRefinerFactoryBase::TopologyDescriptor::TopologyDescriptor() {
-    memset(this, 0, sizeof(TopologyDescriptor));
 }
 
 } // end namespace Far
