+++ tutorials/far/tutorial_4/far_tutorial_4.cpp
@@ -30,7 +30,7 @@
 // factorized stencils to interpolate vertex primvar data buffers.
 //
 
-#include <opensubdiv/far/topologyRefinerFactory.h>
+#include <opensubdiv/far/topologyDescriptor.h>
 #include <opensubdiv/far/stencilTable.h>
 #include <opensubdiv/far/stencilTableFactory.h>
 
@@ -164,7 +164,7 @@ static Far::TopologyRefiner *
 createTopologyRefiner() {
 
     // Populate a topology descriptor with our raw data.
-    typedef Far::TopologyRefinerFactoryBase::TopologyDescriptor Descriptor;
+    typedef Far::TopologyDescriptor Descriptor;
 
     Sdc::SchemeType type = OpenSubdiv::Sdc::SCHEME_CATMARK;
