+++ tutorials/far/tutorial_5/far_tutorial_5.cpp
@@ -31,7 +31,7 @@
 // vertex colors.
 //
 
-#include <opensubdiv/far/topologyRefinerFactory.h>
+#include <opensubdiv/far/topologyDescriptor.h>
 #include <opensubdiv/far/stencilTable.h>
 #include <opensubdiv/far/stencilTableFactory.h>
 
@@ -214,7 +214,7 @@ createTopologyRefiner() {
 
     // Populate a topology descriptor with our raw data.
 
-    typedef Far::TopologyRefinerFactoryBase::TopologyDescriptor Descriptor;
+    typedef Far::TopologyDescriptor Descriptor;
 
     Sdc::SchemeType type = OpenSubdiv::Sdc::SCHEME_CATMARK;
