+++ tutorials/far/tutorial_6/far_tutorial_6.cpp
@@ -40,7 +40,7 @@
 // systems that show the tangent and bi-tangent at the random samples locations.
 //
 
-#include <opensubdiv/far/topologyRefinerFactory.h>
+#include <opensubdiv/far/topologyDescriptor.h>
 #include <opensubdiv/far/primvarRefiner.h>
 #include <opensubdiv/far/patchTableFactory.h>
 #include <opensubdiv/far/patchMap.h>
@@ -272,7 +272,7 @@ static Far::TopologyRefiner *
 createTopologyRefiner() {
 
 
-    typedef Far::TopologyRefinerFactoryBase::TopologyDescriptor Descriptor;
+    typedef Far::TopologyDescriptor Descriptor;
 
     Sdc::SchemeType type = OpenSubdiv::Sdc::SCHEME_CATMARK;
