+++ tutorials/osd/tutorial_0/osd_tutorial_0.cpp
@@ -30,7 +30,7 @@
 // 'Controllers'.
 //
 
-#include <opensubdiv/far/topologyRefinerFactory.h>
+#include <opensubdiv/far/topologyDescriptor.h>
 #include <opensubdiv/far/stencilTableFactory.h>
 #include <opensubdiv/osd/cpuEvaluator.h>
 #include <opensubdiv/osd/cpuVertexBuffer.h>
@@ -138,7 +138,7 @@ createTopologyRefiner(int maxlevel) {
 
     // Populate a topology descriptor with our raw data
 
-    typedef Far::TopologyRefinerFactoryBase::TopologyDescriptor Descriptor;
+    typedef Far::TopologyDescriptor Descriptor;
 
     Sdc::SchemeType type = OpenSubdiv::Sdc::SCHEME_CATMARK;
